Move txncache to flamenco
Fixes a circular dependency between disco and flamenco.
flamenco seems to be the primary user of txncache, so moving it there.
