+++ src/flamenco/runtime/context/fd_exec_slot_ctx.h
@@ -10,7 +10,7 @@
 #include "../sysvar/fd_sysvar_cache.h"
 #include "../sysvar/fd_sysvar_cache_old.h"
 #include "../../types/fd_types.h"
-#include "../../../disco/bank/fd_txncache.h"
+#include "../fd_txncache.h"
 
 struct fd_account_compute_elem {
   fd_pubkey_t key;
@@ -26,7 +26,7 @@ fd_pubkey_eq( fd_pubkey_t const * key1, fd_pubkey_t const * key2 ) {
 
 static ulong
 fd_pubkey_hash( fd_pubkey_t const * key, ulong seed ) {
-  return fd_hash( seed, key->key, sizeof(fd_pubkey_t) ); 
+  return fd_hash( seed, key->key, sizeof(fd_pubkey_t) );
 }
 
 static void
@@ -121,7 +121,7 @@ fd_exec_slot_ctx_recover( fd_exec_slot_ctx_t *   ctx,
 
 /* fd_exec_slot_ctx_recover re-initializes the current slot
    context's status cache from the provided solana slot deltas.
-   Assumes objects in slot deltas were allocated using slot ctx valloc 
+   Assumes objects in slot deltas were allocated using slot ctx valloc
    (U.B. otherwise).
    On return, slot deltas is destroyed.  Returns ctx on success.
    On failure, logs reason for error and returns NULL. */
